Adding messages tests and deleting connection tests file
